Completed react-accordion-component challenge
